created shipyardcontroller_component_test.go
